R program for Estimating Multiple-Trigger Flood Bond Prices
# Multiple-Trigger Flood Bond Pricing Model
This repository contains an R program for estimating Zero-Coupon Flood Bond (ZCFB) and Coupon-Paying Flood Bond (CPFB) prices. The model uses aggregate loss and the maximum number of submerged houses as triggers for flood bonds.

## How to use
Run the script to generate ZCFB and CPFB price estimates based on the input parameters.